requirements.txt
streamlit
pandas
pyyaml
regex
tiktoken
openai
backoff